requirements.txt
numpy
matplotlib
joblib
jupyter
tensorflow==2.10.0
keras==2.10.0
opencv-python
Pillow
